Push

+++ src/main/java/com/mycompany/csc325_oop_designreview_lab/Human.java
@@ -11,10 +11,10 @@ package com.mycompany.csc325_oop_designreview_lab;
 public abstract class Human {
     private String name;
 	private String address;
-	private short age;
+	private int age;
 
 	// constructor that takes only two paras
-	public Human(String name, short age) {
+	public Human(String name, int age) {
 		this.name = name;
 		this.age = age;
 	}
@@ -32,7 +32,7 @@ public abstract class Human {
 	public abstract void setAddress(String address);
 
 
-	public short getAge() {
+	public int getAge() {
 		return age;
 	}
